requirements.txt
discord.py
discord.py[voice]
python-dotenv
yt-dlp
PyNaCl
browser-cookie3